requirements.txt
kafka-python-ng
requests

